requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
diagrams==0.24.4
distlib==0.3.9
distro==1.9.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
fonttools==4.58.5
graphviz==0.20.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.12
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain-core==0.3.61
langgraph==0.4.7
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.1
langgraph-sdk==0.1.70
langsmith==0.3.42
MarkupSafe==3.0.2
matplotlib==3.10.3
neo4j==5.28.1
nodeenv==1.9.1
numpy==2.2.6
openai==1.82.0
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
tenacity==9.1.2
termcolor==3.1.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
virtualenv==20.31.2
Werkzeug==3.1.3
xxhash==3.5.0
zstandard==0.23.0
